refactor: consolidate rocksdb code into submodule
Move all RocksDB checksum code into a separate rocksdb_checksum submodule
that's only compiled when the edge feature is enabled. This reduces the
number of cfg attributes from 10 to just 3.
